feat: write project description & refactor codes
